FPT Project
🛒 Store Inventory and Checkout System

This is a simple text-based Inventory Management and Checkout system for smaller retail environments. The program consists of two roles: Admin and Customer. The Admin and Customer roles are made up of a set of Admin and Customer functionalities dealing with store inventory.

🔧 Features
Admin Menu
Add Product: This allows Admin to add a product into the inventory with unique ID, product name, and price.
Edit Product: This allows Admin to modify the price of an existing product.
Delete Product: This allows Admin to remove a product using the unique ID.
Display Inventory: This allows Admin to view all the products with details about each product.
Switch to Customer Menu: This allows an Admin to switch from the Admin role to a Customer role.
Exit Program

Customer Menu
Search Item: This allows a Customer to search for an item by product name and display details about the product.
Display Most Searched Item: This allows the Customer to view the most searched for product.
Display Inventory: This allows the Customer to view all available products.
Add to Cart: This allows the Customer to add a product to their shopping cart by using product ID and quantity. Checkout: This option calculates the subtotal, adds tax, accepts payment, and concludes the transaction. 

Switch to Admin Menu: This option changes the role from Customer to Admin. 

Exit Program 

🗂️ Data Structures used 

inventory: A list where we store the product entries as [ID, Name, Price]. 

nameCounter: A counter for how many times we add a product name. 

searchCounter: This one is for how many times a product has been searched. 

cart: This list stores the customer cart items as [Product_ID, Quantity]. 

🧠 How it works 

With a simple number menu, an admin can manage the product entries listing. 

Product names from the list are counted to track how often a product has been added (used for quantity logic). 

From an inventory, customers search for product(s), add product(s) to a cart, and checkout. 

Also, when we search for items, we count how many times a product has been searched (to find the most searched product). 

Finally, when we checkout, we add a 13% tax and then validate payment before completing the order.